refactor_clean moving logic to domain package, views to ui package
