+++ app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <com.rusili.lib.parallax.ParallaxThreeLayerView
+    <com.rusili.lib.parallax.ui.ParallaxThreeLayerView
         android:id="@+id/sampleParallax"
         android:layout_width="0dp"
         android:layout_height="256dp"
