+++ lib/build.gradle
@@ -32,5 +32,6 @@ dependencies {
     androidTestImplementation 'androidx.test:runner:1.1.1'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.1'
 
+    testImplementation 'org.amshove.kluent:kluent:1.47'
     testImplementation 'org.amshove.kluent:kluent-android:1.47'
 }
