--- lib/src/test/java/com/rusili/lib/Float3Test.kt
@@ -1,6 +1,0 @@
-package com.rusili.lib
-
-class Float3Test {
-
-
-}
