+++ lib/src/test/java/com/rusili/lib/ParallaxCalculatorTest.kt
@@ -1,6 +1,6 @@
 package com.rusili.lib
 
-import com.rusili.lib.parallax.ParallaxCalculator
+import com.rusili.lib.parallax.domain.ParallaxCalculator
 import org.amshove.kluent.shouldEqual
 import org.junit.Assert.assertEquals
 import org.junit.Test
